+++ src/components/movie-card-full/movie-card-full.test.js
@@ -1,6 +1,78 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
+import {BrowserRouter} from 'react-router-dom';
+import {Provider} from 'react-redux';
+import configureStore from 'redux-mock-store';
 import MovieCardFull from './movie-card-full.jsx';
+
+const mockStore = configureStore([]);
+
+const mockData = {
+  titleMovie: {
+    id: 4,
+    image: `img/we-need-to-talk-about-kevin.jpg`,
+    movieBackground: `img/bg-the-grand-budapest-hotel.jpg`,
+    movieDate: `2000`,
+    movieDescription: `In the 1930s, the Grand Budapest Hotel is a popula…y boy, becomes Gustave&apos;s friend and protege.`,
+    movieDirector: `Wes Anderson`,
+    movieGenre: `Thriller`,
+    movieImage: `img/the-grand-budapest-hotel-poster.jpg`,
+    movieRatingCount: `349 ratings`,
+    movieRatingLevel: `Bad`,
+    movieRatingScore: `3`,
+    movieStarring: [`Jude Law`, `Willem Dafoe`],
+    title: `Dardjeeling Limited`,
+    moviePreview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
+    backgroundColor: `#000000`,
+    movieVideo: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
+    isFavorite: false,
+  },
+
+  moviesList: [
+    {
+      id: 4,
+      image: `img/we-need-to-talk-about-kevin.jpg`,
+      movieBackground: `img/bg-the-grand-budapest-hotel.jpg`,
+      movieDate: `2000`,
+      movieDescription: `In the 1930s, the Grand Budapest Hotel is a popula…y boy, becomes Gustave&apos;s friend and protege.`,
+      movieDirector: `Wes Anderson`,
+      movieGenre: `Thriller`,
+      movieImage: `img/the-grand-budapest-hotel-poster.jpg`,
+      movieRatingCount: `349 ratings`,
+      movieRatingLevel: `Bad`,
+      movieRatingScore: `3`,
+      movieStarring: [`Jude Law`, `Willem Dafoe`],
+      title: `Dardjeeling Limited`,
+      moviePreview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
+      backgroundColor: `#000000`,
+      movieVideo: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
+      isFavorite: false,
+    },
+    {
+      id: 3,
+      image: `img/aviator.jpg`,
+      movieBackground: `img/bg-the-grand-budapest-hotel.jpg`,
+      movieDate: `2008`,
+      movieDescription: `In the 1930s, the Grand Budapest Hotel is a popula…y boy, becomes Gustave&apos;s friend and protege.`,
+      movieDirector: `Christopher Nolan`,
+      movieGenre: `Kids & Family`,
+      movieImage: `img/the-grand-budapest-hotel-poster.jpg`,
+      movieRatingCount: `272 ratings`,
+      movieRatingLevel: `Normal`,
+      movieRatingScore: `2`,
+      movieStarring: [`Jude Law`, `Willem Dafoe`],
+      title: `Johnny English`,
+      moviePreview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
+      backgroundColor: `#000000`,
+      movieVideo: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
+      isFavorite: false,
+    }
+  ],
+  movieGenres: [
+    `Thriller`,
+    `Kids & Family`
+  ]
+};
 
 const movie = {
   id: 4,
@@ -105,16 +177,38 @@ const mockTabs = [
 ];
 
 it(`<MovieCardFull /> should movie full card page to match snapshot`, () => {
+  const {titleMovie, moviesList, movieGenres} = mockData;
+  const store = mockStore({
+    DATA: {
+      titleMovie,
+      moviesList,
+      movieGenres,
+      loadingMovies: false,
+      loadingMoviesError: false,
+    },
+    MOVIE: {
+      activeGenre: `All genres`,
+      moviesRenderLimit: 8,
+    },
+    USER: {
+      authorizationStatus: `AUTH`,
+    }
+  });
   const onMovieTitleClick = jest.fn();
   const tree = renderer
     .create(
-        <MovieCardFull
-          movie={movie}
-          similarMovies={similarMovies}
-          activeItem={`TAB1`}
-          onMovieTitleClick={onMovieTitleClick}
-          tabs={mockTabs}
-        />,
+        <Provider store={store}>
+          <BrowserRouter>
+            <MovieCardFull
+              movie={movie}
+              similarMovies={similarMovies}
+              activeItem={`TAB1`}
+              onMovieTitleClick={onMovieTitleClick}
+              tabs={mockTabs}
+              history={{}}
+            />
+          </BrowserRouter>
+        </Provider>,
         {
           createNodeMock: () => {
             return {};
