+++ src/reducer/data/data.test.js
@@ -80,6 +80,7 @@ describe(`Reducer works correctly`, () => {
   it(`Reducer without additional parameters should return initial state`, () => {
     expect(reducer(void 0, {})).toEqual({
       loadingMovies: false,
+      loadingMoviesError: false,
       loadingComments: false,
       loadingTitleMovie: false,
       moviesList: [],
