+++ src/reducer/data/selectors.js
@@ -30,7 +30,7 @@ export const getTitleMovieLoadingStatus = (state) => {
   return state[NAME_SPACE].loadingTitleMovie;
 };
 
-export const getMovieById = (state, id) => {
+export const getMovieByIdFromState = (state, id) => {
   const [movie] = state[NAME_SPACE].filter((it) => it.id === id);
   return movie;
 };
